Lots of cleanup and fixes. Now getting quite far in rom code. Implemented eXecute for branch/jump with subject execute. Getting to the code trying to setup the ROM and RAM spec regs, some Calculate instruction is leaving a bit set or something.

+++ mmu.h
@@ -63,7 +63,7 @@ uint32_t procread (uint32_t addr, uint8_t bytes, uint8_t mode);
  * I/O Address Assignments pg. 11-136
  */
 union MMUIOspace {
-	uint8_t _direct[MMUCONFIGSIZE];
+	uint32_t _direct[MMUCONFIGSIZE];
 	struct {
 		uint32_t SegmentRegs[16];
 		uint32_t IOBaseAddr;
